+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Updates
+ - changed client error handling to display docstrings on bad input
+
 ## [0.2.31] - 2024-07-29
 
 ### Added